Implement EBS volume mounting
This can either be invoked by instrumenting the user-data with a mime
part with content-type 'text/x-ebs-mount-description' with a body like
so:

device=/dev/sde:/var/lib/mysql,/etc/alfresco
device=/dev/sdf:/other/things

or by using the appliance config XML format like so:

<appliance>
  <storage device="/dev/sde">
	<path>/var/lib/mysql</path>
    <path>/etc/alfresco</path>
	</storage>
  <storage device="/dev/sdf">
    <path>/other/things</path>
  </appliance>
</appliance>

In either case, if the volume does not yet have a filesystem, one will be created.

For each path that is to live on the volume, a directory is created, and
populated with the data currently in the target directory (e.g.
/var/lib/mysql is copied to ${ebs_volume_path}/_var_lib_mysql). Once
this is done, the directories are bind-mounted to the relevant paths.
If the directories in question already exist, they will just be bind-mounted.