functional cloud-config-ssh.conf now
This fixes LP: #506599 , LP: #507070 .

Also, here, we move config parsing out of the EC2Init class.
Instead the parsing of /etc/ec2-init/ec2-init.cfg is now done in the
CloudConfig class, and prepended to the cloud-config that comes in
from user data.  This means that values in user data will over-ride
the settings in /etc/ec2-init.

Still TODO is support reading the public and private key files from
the cloud-config