add 'cloud-boothook' type
if user data is of type text/cloud-boothook, or begins with
#cloud-boothook, then assume it to be code to be executed.

Boothooks are a very simple format.  Basically, its a one line header
('#cloud-config\n') and then executable payload.

The executable payload is written to a file, then that file is executed
at the time it is read.  The file is left in 
/var/lib/cloud/data/boothooks

There is no "first-time-only" protection.  If running only once is
desired, the boothook must handle that itself.