fix bug where nfs/network mounts could not be specified (LP: #603329)
What caused this was having an open ended list on what "names" might be
found in the metadata service.  That list is now trimmed down to the
the following values:
 ephemeral*
 root
 ami
 swap

The above list was found from crawled medata data services in the latest
maverick tests I did.  The following is the complete list of entries that
were there:
 'ami': '/dev/sda1',
 'ami': 'sda1',
 'ephemeral0': '/dev/sda2'
 'ephemeral0': '/dev/sdb'
 'ephemeral0': 'sda2'
 'ephemeral0': 'sdb'
 'ephemeral1': 'sdc'
 'ephemeral2': 'sdd'
 'ephemeral3': 'sde'
 'root': '/dev/sda1'
 'root': '/dev/sda1'
 'swap': 'sda3'

Also, this limits which devices will have "/dev/" prepended to them to
sda, sda1, xvda, xvda1, hda1, hda, vda.