write the previous-hostname file to persistent cloud/data
This change just uses a different facility for coming up with the path.
But, by design I'm chosing to put 'previous-hostname' in 
 /var/lib/cloud/data/
rather than in
 /var/lib/cloud/instance/data/

The idea is that if the user:
 - started an instance
 - modified /etc/hostname
 - bundled instance (or create-image from it)
 - started new instance

They would expect their modified /etc/hostname to persist.  As such, the
previous-hostname file should be cross-instance data.

Bugs in this area include LP: #596993 and LP: #514492