fix cloud-run-user-script for directory change, make it invoke other scripts
Now, in addition to running instance specific scripts (in runcmd or
user-data scripts), cloud-run-user-script will run other directories
also.  All under /var/lib/cloud, and in the following order:
   scripts/per-once [once ever]
   scripts/per-boot [every boot]
   scripts/per-instance [once per instance]
   instance/scripts [once per instance]

At the moment, the marker is on the entire directory, so changes to that
directory. Changes to the contents of the directory will not be noticed.