handle no datasource better cloud-init-cfg
previously, cloud-init-cfg would call CloudConfig with no cloud
specified.  Then, CloudConfig would get a new cloudinit instance
and specify no ds_deps (which means network and local).
If no cache was found, it would wait again for the Ec2 metadata
service (even if it wasn't supposed to).

Now, we get the cloud instance in the main program and attempt to
get the data source, but only from [], which will get only cached.

Then, if its not found, exit silently.