increase timeout on read_seed when a seedfrom was explicitly given
In the case where a seedfrom value was given on the command line or in the
config file, we were timing out in 2 seconds on the connection.  That timeout
was put in place to support "probing" for sources, but seedfrom is explictly
given.

So, in that case, do a urllib.open without a timeout value.  Looking at source
code, default timeout is 'socket._GLOBAL_DEFAULT_TIMEOUT', but rather than
importing that and using  it, I will call without a timeout value.

LP: #812646