This fixes LP: #819507, to make consume_userdata run 'always'
consume_userdata should really run always, rather than once per instance.

The documentation says that boothooks were on their own for per-instance
but since this routine was only being called once, they would only get
called once.

This  modifies the behavior to be:
   user_script: per_always
   cloud_config : per_always
   upstart_job : per_instance
   cloud_boothook: per_always

In order to not break part handlers that are existing, and expect to only be
called once per instance, this adds a 'handler_version' item in a handler
that can indicate the version (currently 1 or 2).  If it is 2, then the
hander will be passed the frequency (per-instance or per-always) that this
is being run.  That way the handler can differenciate between them.

This also makes 'bootcmd' run every boot.  That should be changable in
cloud-config though, so users who dont like the behavior can modify it.

LP: #819507