make DataSourceEc2 more resilliant to slow metadata service (LP: #894279)
This increases the timeout for a metadata request to something that should
be easily satisfiable (50 seconds).  But hopefully does so while still keeping
the case of no-metadata service in mind.

Previously, there was a small timeout and many retries (30) would be done.
Now,
 - larger timeout (50 seconds) by default
 - retry until a given "max_wait" is reached (120 seconds default)

The end result is that if we're hitting the timeout, there will only end up
being a couple attempts made.  But if the requests are coming back quickly
then we'll still make several attempts.

There is one EC2DataSource config change, now 'retries' is not used, but rather
'max_wait' to indicate generally how long it should try to find a metadata
service.