revert default handling of /etc/hosts to 0.6.1 style (Ubuntu 11.04)
The default management of /etc/hosts in 0.6.2 (Ubuntu 11.10)
was problematic for a couple different uses, and represented a change
in what was present in previous releases.

This changes the default behavior back to the way it was in 11.04/0.6.1.

It makes 'manage_etc_hosts' in cloud-config more than just a boolean.
It can now have 3 values:

 * False (default): do not update /etc/hosts ever
 * "localhost": manage /etc/hosts' 127.0.1.1 entry (the way it was done
   in 11.10/0.6.2)
 * True (or "template"): manage /etc/hosts via template file

This addresses bugs 
  * LP: #890501 
  * LP: #871966