run resizefs module on every boot
on EC2, you can:
 stop instance
 resize root volume
 start instance

Currently, the partition would get grown correctly in the initramfs, but
the root filesystem will not get automatically resized in that case as it
only runs per_instance.

This should not be harmfull in any case, as resizefs will just report
nothing to do:

$ sudo resize2fs /dev/sda5
resize2fs 1.42-WIP (16-Oct-2011)
The filesystem is already 25600278 blocks long.  Nothing to do!