better 'main()', add and use version in api, do not require user-data
main now is more useful for debugging. now it does:
 * get: just dump contents of a url provided after oauth
 * crawl: walk through using indexes, dumping content
 * check-seed: validate the seed is good

uses MD_VERSION in the url, and appends that to the metadata url
in the config file. (previously it assumed the url in the config
was the full url)

does not require user-data in the http seed.  if the user did not
specify user-data, it wont be there, so do not fail on that case.