do not disable logging on signal
I had previously asked for this, but we're hoping to handle it in a
more generic way.  Just because we receive a signal doesn't mean that
all logging is broken.

The more general solution we'll chase is to catch a failure of a log
message and fall back if necessary across the board.  That way cloud-init
will still send logging to the right places on a user interupt.