send stderr from write-ssh-key-fingerprints to stdout
This changes all output write-ssh-key-fingerprints to go to its
stdout by redirecting stderr to stdout. 

The reason for this is that cc_keys_to_console.py was swallowing stderr
and not replaying it to /dev/console.

Ideally, we'd have a way in 'util.subp' to do effectively the same thing
as we're doing here in the shell script.

LP: #1055688