fix oauth time skew. actual implementation was returning 401 not 403.
This fixes (tested) bug 978127.  The server was actually returning a 401
not a 403.  As such, the fix here was insufficient.  This will now take
either of those 2 error codes.  I've also tested it by changing the clock
in the cloud-init upstart job with a stanza like below, and verifying
that we do see the problem and then it resolve itself:

  pre-start script
    offset="10 minutes ago"
    past=$(date -R --date "$offset")
    date --set "$past" &&
      echo ===== "set date to $past [$offset]" ===== ||
      echo ===== "failed to set date to $past [$offset]" ====
  end script

LP: #978127