change the "null datasource" for stages.py to None
Previous commit set up an object for checking "None" in the stages class.
Here, change that to be None.  The benefit of the DataSourceNone was so that
the stages was generally functional even with that "None" object.

The negative affect of that is that you could go down the error path much
longer when using it.  If we use None type, then self.datasource.<something>
will fail immediately and loudly.