ensure a datasource's 'distro' and sys_cfg are updated
After parsing and merging datasource's config, the changes in were not making
it into the datasource's 'distro.

The end result was that the when a config module was called, it's 'cloud'
argument would be updated in 'cloud.distro', but not in
'cloud.datasource.distro'.  This path was required for getting mirror settings
to take affect, because they include information from the datasource.

Ie:
  cc_apt_configure had
    mirror_info = cloud.datasource.get_package_mirror_info()
  the datasource then used *its* copy of sys_cfg to call
    self.distro.get_package_mirror_info
  and *that* distro's sys_cfg had not been updated.

LP: #1090482