DataSourceConfigDrive: consider CD rom as valid config-drive source.
previously, there was an attempt in the config drive source to limit
the source device to a "full block device" rather than a partition.

This was done by a simplistic approach of checking that the last 
character of the name was not a number.  That was filtering out
CD-rom devices (sr0).

Now, we have a bit more sophisticated approach to that same problem.
We filter out block devices that have a 'partition' entry in
/sys/class/block/DEVICE_NAME/partition .