fix DataSource base class to set up ds_cfg for 'Net' sources
When the base DataSource class would set 'ds_cfg' for the specific
datasources' config, it would fail for the DataSources that are just named
'DataSourceFooNet' and we wanted to set configuration in 'Foo'.

For example, both DataSourceOpenNebula and DataSourceOpenNebulaNet want to
read datasource config from
 sources:
  OpenNebula:
   foo: bar

But without this change, 'ds_cfg' would not be setup properly for
OpenNebulaNet.