cc_growpart: respect /etc/growroot-disabled
cloud-initramfs-growroot is an initramfs module in cloud-initramfs-tools
that resizes the root partition before the root is pivoted over.

growroot was used in Ubuntu up to and including 12.10.  The file
/etc/growroot-disabled on the root filesystem was the only way of
disabling the growing of the root partition.

In cloud-init 0.7.2 cloud-init began resizing the root partition
as growpart gained the ability to utilize 'ptupdate' in kernels > 3.8.
This was a big improvement as now the user could disable or enable
the growing of the root partition via user-data.

In order to let users disable growing of / very simplistically cloud-init
will now respect the presense of /etc/growroot-disabled unless config
specifically tells it to ignore that file.

LP: #1234331