support calling apt with eatmydata, enable by default if available.
This allows a general config option to prefix apt-get commands via
'apt_get_wrapper'.  By default, the command is set to 'eatmydata', and the
mode set to 'auto'.  That means if eatmydata is available (via which), it
will use it.

The 'command' can be either a array or a string.