Add 'unverified_modules' config option and skip unverified modules
Config modules are able to declare distros that they were verified
to run on by setting 'distros' as a list in the config module.

Previously, if a module was configured to run and the running distro was not
listed as supported, it would run anyway, and a warning would be written.

Now, we change the behavior to skip those modules.

The distro (or user) can specify that a given list of modules should run anyway
by declaring the 'unverified_modules' config variable.

run_once modules will be run without this filter (ie, expecting that the user
explicitly wanted to run it).