change 'default' encoding to be "None"
Instead of just trying to see if userdata decodes as the indication that
it should be encoded, the user must explicitly set this.

The "just try it" will fail in the case where the user had other use
of user-data and wanted a blob of data to go through unrecognized by
cloud-init.

In cases where there can be mistake in automatic behavior,
and some users may be relaying on old behavior, its best to just require
explicit use.