Add the first layout of the new osys architecture
This patch provides some of the base classes, which composes the new
operating system utils architecture, based on a OOP approach and separation of
concerns. The utils is namespace is composed from other subnamespaces,
each dealing with one particular area, such as networking or filesystem interaction.
A first layout for the network namespace is provided, as well as a first
operating system utils namespace for the Windows platform.