events: add timestamp and origin, support file posting
This adds 'timestamp' and 'origin' to events.
The timestamp is simply that, a floating point timestamp of when
the event occurred.

The origin indicates the source / reporter of this.  It is useful
to have a single endpoint with multiple different things reporting
to it.  For example, MAAS will configure cloud-init and curtin
to report to the same endpoint and then it can differenciate who
made the post.  Admittedly, they could use multiple endpoints, but
this this seems sane.

Also, add support for posting files at the close of an event.
This is utilized in curtin to post a log file when the install is
done.  files are posted on success or fail of the event.