No longer run pollinate by default in seed_random
The user can still choose to run pollinate here to seed their
random data.  And in an environment with network datasource, that
would be expected to work.  However, we do not want to run it any 
more from cloud-init because
a.) pollinate's own init system jobs should get it ran before ssh,
    which is the primary purpose of wanting cloud-init to run it.
b.) with a local datasource, there is no network guarantee when
    init_modules run, so pollinate -q would often cause issues then.
c.) cloud-init would run pollinate and log the failure causing
    many cloud-init specific failures that it could do nothing about.

LP: #1554152