quickly check to see if the previous instance id is still valid
This adds a check in cloud-init to see if the existing (cached)
datasource is still valid.  It relies on support from the Datasource
to implement 'check_instance_id'.  That method should quickly determine
(if possible) if the instance id found in the datasource is still valid.

This means that we can still notice new instance ids without
depending on a network datasource on every boot.

I've also implemented check_instance_id for the superclass and for
3 classes:
  DataSourceAzure (check dmi data)
  DataSourceOpenstack (check dmi data)
  DataSourceNocloud (check the seeded data or kernel command line)


LP: #1553815