make NoCloud work for seeding network.
Tested now with the generated fallback config in an lxc container.
Had to change to return a config rather than a network state.

Also this makes nocloud look in nocloud-net's seed dir.
This way it will read the seed and clame the datasource but
not do anything other than apply networking and the init_modules early.

It is a change in behavior of the time that boothooks woudl run to do
this.  May need to change that back.