trust existing datasource in modules or single
This fixes a bug where modules mode was not passing a 'existing'
flag to fetch.  fetch had existing default to 'check'.
The DataSourceNoCloud when fed with data from a disk will return
False to check() as it is not a guarantee'd hit.

That caused fetch to go looking for a new datasource.

That would have actually worked, but modules and single create
the Init with deps=[].  So it went looking for Datasources
that matched those deps, and only found DataSourceNone.

I'm going to keep having modules and single specify deps=[]
as that will prevent them from going to look for a DS and
further making things worse.