improve how cloud-init-wait waits
since we run from a udev event, this is sufficient.
udev settle will exit when either of
 a.) the file exists
 b.) the udev event queue has all been processed.

Since cloud-init-wait is being run as a udev event, 'b' cannot
be satisfied until it finishes.  Thus, this essentially becomes a
inotify based wait for the file /run/cloud-init/network-config-ready
and no loops are needed.