chef: straighten out validation_cert and validation_key
Now, validation_key is always a path to a file, as it is in 
chef's client.rb syntax.

validation_cert is always the *content* of that file that should
be written.  However, if validation_cert is the string "system",
then we do not write that value, but rather assume the file exists.

LP: #1568940