support and render control=manual on initramfs network devices
when reading the initramfs configurewd devices and turning them
into network config, we change to not have 'auto' control (or allow=auto).
The reason for this is that if the device was still up:
 a.) it would try to bring it up again (due to bug 1570142)
 b.) it would be brought down.

'b' is problematic if there is an iscsi or network root filesystem.

Note, that ifupdown does now support 'no-auto-down' which means
that the nic should not be brought down on 'ifdown -a'.

LP: #1568637