net: fix inet value for subnets, don't add interface attributes to alias
[copied from curtin revno 390]
Apply two separate fixes for configuring bonding with ip aliases.

Curtin re-used the interface's inet value for each subnet that might
be configured.  In the case where the configuration included an ipv4
address after an ipv6 one resulted in emitting 'inet6' for ipv4 address
which is not correct.  Resolve this issue by calculating the inet
value independent of the current status of the iface, using the subnet
config instead.

When rendering a network_config which includes ip alias interfaces
do not emit any attributes, like MTU, or bond/bridge options  Including
these values is almost always wrong or will result in confusing
behavior on the target system.

LP: #1588547