Just use file logging by default
Instead of being dependent on the availability of syslog that
various distributions may not enable or configure correctly
or they do so via patches just use a known-to-work default
logging mechanism.  If distros want to change this, that
is fine, but at least the built-in one will work reliably.

LP: #1643990