Generate a dummy bond name for OpenStack
The OpenStack network_data.json does not provide a name for bond links.
This change makes it so a dummy one is generated and used instead
to satisfy cloud-init which does require one.

In order to write the correct link (underlying 'link' names)
for the bonds, we maintain a list of info by ids so we can easily
get the right device name.

Also:
  * add a vlan test case that similarly references an id rather than name.
  * make bond interfaces auto

LP: #1605749