drop modification of version during make-tarball, tools changes.
Modification of the tarball became problematic, as it meant that
any tool extracting source would find the orig source tarball different.
I found this unusable when trying to use 'gbp buildpackage'.

Other changes here are to better support using python3 or python2
for the build.  Makefile will try to call the right python version
and can be told which python to use.

read-version: by adding 'tiny_p' and avoiding the import of
cloudinit.util, we need less dependencies to run this.