azure dhclient-hook cleanups
This adds some function to the generator to maintain the presense of a
flag file '/run/cloud-init/enabled' indicating that cloud-init is enabled.

Then, only run the dhclient hooks if on Azure and cloud-init is enabled.
The test for is_azure currently only checks to see that the board vendor
is Microsoft, not actually that we are on azure.  Running should not be
harmful anywhere, other than slowing down dhclient.

The value of this additional code is that then dhclient having run
does not task the system with the load of cloud-init.

Additionally, some changes to config are done here.
 * rename 'dhclient_leases' to 'dhclient_lease_file'
 * move that to the datasource config (datasource/Azure/dhclient_lease_file)

Also, it removes the config in config/cloud.cfg that set agent_command
to __builtin__.  This means that by default cloud-init still needs
the agent installed.  The suggested follow-on improvement is to
use __builtin__ if there is no walinux-agent installed.