network: fix get_interface_mac for bond slave, read_sys_net for ENOTDIR
When using get_interface_mac, on a system with bond slaves, it would
return the bond_master's address.  That isn't expected, and causes
problems in a caller like get_interfaces_by_mac which would then seem to
find duplicate macs on the system.

Additionally, in read_sys_net catch a errno.ENOTDIR error as ENOENT.

Opening a path as a file that has <existing_file>/anything will will raise
ENOTDIR rather than ENOENT.  This handles that case in read_sys_net as a
if the file did not exist.