DataSourceOVF: fix user-data as base64 with python3
When user-data was provided in the ovf environment python3 would call
base64.decodestring() with a string rather than bytes and an exception
would occur.

This fixes the broken path and adds unit test.  Also changes to
return None rather than empty string when there is no user-data and
when there is user-data return that as bytes instead of string.

LP: #1619394