lxd: Update network config for LXD 2.3
Prior to LXD 2.3, the bridge configuration was done through distro
packaging.  Thus, lxd module interacted with debconf.
With 2.3 and higher, this is now done inside LXD itself, so we
need to use "lxc network" there.

For now, this perfectly matches what we had before with debconf and
doesn't cover any of the new options. We can always add those later.

A set of tests similar to what we had for debconf has been added to make
sure things look good.

This is tested in Yakkety container running LXD 2.3 and all options seem
to be passed through as expected, giving me the bridge I defined.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>