tests: silence the Cheetah UserWarning about NameMapper C version.
This silences a warning made by Cheetah in pip installed environments:
  UserWarning: You don't have the C version of NameMapper installed!
  I'm disabling Cheetah's useStackFrames option ...

The reason for the monkey patching is that the warning goes to stderr
during nose and breaks up its expected output.  The side affect of it
is that tests would run with Cheetah's 'useStackFrames'
enabled which is "painfully slow with the Python version of NameMapper".