Move user/group functions to new ug_util file
The amount of code to do user and group normalization
and extraction deserves its own file so move the code
that does this to a new file and update references to the
old location.

This removes some of the funkyness done in config modules
to avoid namespace and attribute clashes as well.