systemd: run cloud-init.service Before dbus.service
This is the best work around we have for an issue seen when dns is used
during cloud-init.service on a system that is configured in nsswitch
to use systemd-resolved.  The problem is that cloud-init.service was
blocking basic.target, and dbus would not be available until after
cloud-init.service finished.

LP: #1629797