update Gentoo initscripts to run in the correct order
cloud-init-local needs to be run before net (and in the boot runlevel)
as it sets up networking config and adds it to the default runlevel.

cloud-init-local needs to be run in the boot runlevel because it
modifies services in the default runlevel.  When a runlevel is started
it is cached, so modifications that happen to the current runlevel while
you are in it are not acted upon.

cloud-init needs to run after net, it does not need net because we still
want it to set up other things if it can.