systemd: networking service adjustments.
Here we drop the Requires=networking.service but keep the
After=networking.service.  The change there is that we no longer
force networking.service to run, as the system may not have it or
be configured to use it.

Second, we add After=systemd-networkd-wait-online.service.
That causes us to run after systemd-networkd has brought networking
up.

The end result is that we now run after both ifupdown or
systemd-networkd (if they were going to run) but do not force either
to run.

One other change is that by dropping Requires=networking.service
cloud-init will now run even if networking failed to come up.

LP: #1636912