systemd: cloud-init remove After=systemd-networkd-wait-online
cloud-init.service cannot currently run
  After=systemd-networkd-wait-online.service
as systemd-networkd.service is After dbus.service, but
cloud-init.service is:
  Before=sysinit.target
and sysinit.target is before dbus.service.

The result is that we are temporarily preferring for cases
where there is systemd-networkd in place to have cloud-init.service run
without networking properly configured.

See bug 1636912 for more information.