Add activate_datasource, for datasource specific code paths.
This adds a call to 'activate_datasource'.  That will be called
during init stage (or init-local in the event of a 'local' dsmode).
It is present so that the datasource can do platform specific operations
that may be necessary.  It is passed the fully rendered cloud-config
and whether or not the instance is a new instance.

The Azure datasource uses this to address formatting of the ephemeral
devices.  It does so by
 a.) waiting for the device to come online
 b.) removing the marker files for the disk_setup and mounts modules
     if it finds that the ephemeral device has been reset.

LP: #1611074