main: set TZ in environment if not already set.
If TZ is not set in environment, then datetime.datetime() will
stat /etc/localtime on every usage.  datetime is used in the logging
path of cloud-init, and as such this is very busy.
During a normal cloud-init boot, it results in we roughly
400+ stat calls to /etc/localtime.

The value we set is :/etc/localtime, which is described at
  https://www.gnu.org/software/libc/manual/html_node/TZ-Variable.html

See also merge proposal at 307722 for more background.
https://code.launchpad.net/~raharper/cloud-init/+git/cloud-init/+merge/307722