when adding a user, strip whitespace from group list
The documentation shows group names in the 'groups:' key delimited by
", ", but this will result in group names that contain spaces.  This
can cause the 'groupadd' or 'useradd' commands to fail.

This patch ensures that we strip whitespace from either end of the
group names passed to the 'groups:' key.

LP: #1354694