set_hostname: avoid erroneously appending domain to fqdn
In some situations, cloud-init will erroneously append a default
domain to an already fully qualified hostname, resulting in something
like 'localhost.localdomain.localdomain'.  This patch checks to see if
the value returned by util.get_hostname() contains a '.', and if it
does treats it as a fully qualified name.

Resolves: rhbz#1389048
LP: #1647910