doc: change 'nobootwait' to 'nofail' in docs
'nobootwait' is an upstart specific extension to the mount syntax that is
not supported by other mount systems.  As Ubuntu 16.04 moved from upstart
to systemd, support for 'nobootwait' was lost.

All examples using 'nobootwait' are updated to use the standard 'nofail',
which gives the expected behaviour of not failing to boot in case a volume
is missing. There are subtle differences in semantics between
'nobootwait' and 'nofail', but it is the best substitute that gives
behaviour similar to the upstart specific option.