user-groups: fix bug when groups was provided as string and had spaces
Cloud-config provided like:
  users:
    - default
    - name: foobar
      groups: sudo, adm

Would result in adduser being called as:
  useradd foobar --groups 'sudo, adm' -m
Which would cause error:
  useradd: group ' adm' does not exist

The fix here is just to always normalize groups and remove whitespace.

Additionally a fix and unit tests to explicitly set system=False
or no_create_home=True.  Previously those paths did not test the value
of the entry, only the presense of the entry.

LP: #1354694