doc: Fix typos and clarify some aspects of the part-handler
The existing documentation referred to a handle_type method when it
really should be handle_part.  It also referred to 'methods' when it
really should say 'functions' to be clear (while it's true the built-in
handlers are classes with methods of these names, in this context we
mean module-level functions).

Also clarified that a part-handler should come before the parts that
it handles, and can override built-in handlers.