Add tools/ds-identify to identify datasources available.
ds-identify is run here from the generator.  If ds-identify does
not see any datasources, it can completely disable cloud-init.
The big value in this is that if there is no datasource, no python
will ever be loaded, and cloud-init will be disabled.o

The default policy being added here is:
  search,found=all,maybe=all,notfound=disabled

That means:
  - enable (in 'datasource_list') all sources that are found.
  - if none are found, enable all 'maybe'.
  - if no maybe are found, then disable cloud-init.

On platforms without DMI (everything except for aarch64 and x86),
the default 'notfound' setting is 'enabled'.  This is because many of
the detection mechanisms rely on dmi data, which is present only on
x86 and aarch64.