ec2_utils: fix MetadataLeafDecoder that returned bytes on empty
the MetadataLeafDecoder would return a bytes value b'' instead of
an empty string if the value of a key was empty.  In all other cases
the value would be a string.

This was discovered when trying to json.dumps(get_instance_metadata())
on a recent OpenStack, where the value of 'public-ipv4' was empty.
The attempt to dump that with json would raise
 TypeError: b'' is not JSON serializable