ds-identify: only run once per boot unless --force is given.
This makes ds-identify run only once.  Previously it would run multiple
times each boot as the generator would run more than once.  This is
potentially dangerous, in that running again might find more attached
disks.  However that is really only a "lucky" fix if it happens to
result differently than the first run.

Additionally, we now log the uptime that we started and ended at.