tools/ds-identify: disable vmware_guest_customization by default.
ovf_vmware_guest_customization defaults to true in cloud-init, meaning
that such customization is disabled.  We just missed a return value
causing ovf_vmware_guest_customization to effectively default to on.

Also, when looking for setting look at /etc/cloud/cloud.cfg.
This had been omitted in interest of performance, but we should
be looking there.