Z99-cloud-locale-test.sh: install and make consistent.
Modify upstream packaging to install this file, which was
already installed in ubuntu packaging.

Also, white space changes from tabs to spaces.  Very few things
in cloud-init are tabs now.

Lastly, remove the executable bit on this as ait is not necessary.
Scripts in /etc/profile.d do not have executable bit.