ds-identify: report cleanups for config and exit value.
Change policy so that 'report' can be overridden.
In xenial we had set the builtin default to be 'report', expecting
that Ubuntu core would install config that changed it to 'search'.
However, if report was already set, there was no way to unset it.

The change here is to make 'report' basically 'search-dryrun', so
that one or the other can be set.

The other change here is that report would actually exit disabled
if it did not find a datasource and notfound=disabled.  That was
unexpected and would turn cloud-init off, which is not what we wanted.

Additionally, consistently use 'enabled' or 'disabled' versus
'enable' and 'disable'.

LP: #1669949