GCE: Search GCE in ds-identify, consider serial number in check.
While documentation indicates that the smbios product name should
contain 'Google Compute Engine', experimentation and bug reports
indicate that is not always the case.  The change here is to change
the check for GCE to also consider a serial number that starts with
'GoogleCompute-'.

Also, ds-identify was not currently searching for GCE if no config of
datasource_list was found.  Most images have a datasource_list defined.
So update the list to include GCE.

LP: #1674861