DigitalOcean: assign IPv4ll address to lowest indexed interface.
Previously the IPv4LL address for metadata discovery was assigned to the
first interfaces from an alphabetic sort. On DigitalOcean, the metadata
is only accessible from the first interface. This fixes a problem where the
IPv4LL address is bound to the wrong interface with snapshots.

This is part of general improvements to the DigitalOcean Datasource in
bug 1676908.