net: Allow netinfo subprocesses to return 0 or 1.
On systems with selinux enabled, some of the networking commands executed
successfully do not return 0.  Allow these commands to return 1 since the
output is valid.

Ultimately we need to get this information in some way so that we can
display it correctly.  For now, work around the stack trace when selinux
does not allow us to collect it.

LP: #1686751