fs_setup: if cmd is specified, use shell interpretation.
If 'cmd' is provided to a fs_setup entry, then cloud-init was trying
to execute the rendered string as a single name, rather than
splitting the string.  The change here will pass the string to
shell for interpretation so that it is split there.

Also fix some documentation errors and warn when fs_opts or overwrite
is provided along with 'cmd'.

LP: #1687712