net: when selecting a network device, use natural sort order
The code deciding which interface to choose as the default to request the
IP address through DHCP does not sort the interfaces correctly. On Ubuntu
Xenial images for example, the interfaces are named ens1, ens2, ens3...,
ens11, ... depending on the pci bus address. The python sorting will list
'ens11' before 'ens3' for example despite the fact that 'ens3' should be
before 'ens11'.

This patch address this issue and sort the interface names according to a
human sorting.

Signed-off-by: Marc-Aurèle Brothier <m@brothier.org>